backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain-community>=0.3.21",
    "langchain-ollama>=0.3.3",
    "langchain-postgres>=0.0.14",
    "psycopg>=3.2.9",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "ulid-py>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
testpaths = ["tests"]
